mnist/backend/requirements.txt
flask
Pillow==9.1.0
torch
torchvision
